+++ src/app/admin/emails/page.tsx
@@ -1,0 +1,5 @@
+import { EmailsView } from "@/components/admin/views/EmailsView";
+
+export default function AdminEmailsPage() {
+  return <EmailsView />;
+}